handle refs to builtin types and unions correctly
