+++ src/Codegen/Typing/TypeAlias.hack
@@ -24,6 +24,11 @@ final class TypeAlias extends Type {
   }
 
   <<__Override>>
+  public function getAlias(): string {
+    return $this->alias;
+  }
+
+  <<__Override>>
   public function hasAlias(): bool {
     return $this->getType()->hasAlias();
   }
@@ -35,6 +40,6 @@ final class TypeAlias extends Type {
 
   <<__Override>>
   public function render(): string {
-    return $this->alias;
+    return $this->getAlias();
   }
 }
